% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls_exp.R
\name{xlsx_exp}
\alias{xlsx_exp}
\title{Exporting data into .xlsx files}
\usage{
xlsx_exp(
  data,
  sheetnames = NULL,
  filename,
  path,
  keepNA = F,
  colnames = T,
  rownames = F
)
}
\arguments{
\item{data}{data frame or list of data frames}

\item{sheetnames}{character vector used to name each file sheet}

\item{filename}{character to name the exported file}

\item{path}{path where to store the file inside the working directory}

\item{keepNA}{logical, indicating if missing values should be displayed or
not. If TRUE, are displayed. Default set to FALSE. Note: if keepNA = TRUE,
and the column is numeric, it would turn into character.}

\item{colnames}{logical, indicating if column names should be in the output. If
true, column names are in the first row. Default set to TRUE.}
}
\value{
.xlsx file exported in the indicated path.
}
\description{
Exporting one or multiple tables into a .xlsx file.
If there are multiple tables, each one will be stored in the same file in
different sheets.
}
